Fix data type

+++ WeChatFerry/spy/chatroom_mgmt.cpp
@@ -10,7 +10,7 @@
 using namespace std;
 
 extern WxCalls_t g_WxCalls;
-extern DWORD g_WeChatWinDllAddr;
+extern UINT64 g_WeChatWinDllAddr;
 #if 0
 int AddChatroomMember(string roomid, string wxids)
 {
