+++ WeChatFerry/spy/receive_transfer.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 extern WxCalls_t g_WxCalls;
-extern DWORD g_WeChatWinDllAddr;
+extern UINT64 g_WeChatWinDllAddr;
 #if 0
 int ReceiveTransfer(string wxid, string transferid, string transactionid)
 {
